requirements.txt
PyQt6
PyYAML
cryptography
